Data.Foldable.Free

Description

This module provides a definition of a free Foldable container in terms of Fold, using existential types.

Data types

newtype FoldableFree a

A free Foldable container

Constructors

FoldableFree

Fields

runFoldableFree :: forall x. Fold a x -> x

Construction/Destruction

head :: FoldableFree a -> Maybe a

The first element of a FoldableFree, if it exists

tail :: FoldableFree a -> FoldableFree a

All but the first element of a FoldableFree if it's non-null. Otherwise, returns empty

uncons :: FoldableFree a -> Maybe (a, FoldableFree a)

Return Just the head and tail or Nothing if null

cons :: a -> FoldableFree a -> FoldableFree a

Prepend an element to a FoldableFree

Failing implementations:

 cons x FoldableFree {..} = runFoldableFree $ stepFold x foldableFree

 cons' x FoldableFree{..} = runFoldableFree $
   Fold (flip cons') (return x) id

 λ> cons' () [()]
 FoldableFree *** Exception: stack overflow

Conversion

foldableFree :: Fold a (FoldableFree a)

Fold into a FoldableFree

toFoldableFree :: Foldable t => t a -> FoldableFree a

Convert any Foldable into a FoldableFree.

 toList . toFoldableFree == toList

Distributive/Representable

distributeFold :: Functor f => f (Fold a b) -> Fold a (f b)

Uses tabulateFold and indexFold

foldR :: Foldable f => Fold a b -> f a -> b

Apply a strict left Fold to a Foldable container, right-to-left

indexFold :: Fold a b -> FoldableFree a -> b

We build tabulate by folding over the input values and feeding them to the FoldableFree.

 index == foldR

 instance Representable (Fold a) where
   type Rep (Fold a) = FoldableFree a

tabulateFold :: (FoldableFree a -> b) -> Fold a b

Tabulate a Fold by building up successive FoldableFrees.

Quick benchmark:

 λ> F.fold ((tabulateFold . indexFold) (F.lastN 3)) ([1..2^12] :: [Int])
 [4094,4095,4096]
 (1.67 secs, 1,952,032,256 bytes)

 λ> F.fold (F.lastN 3) ([1..2^12] :: [Int])
 [4094,4095,4096]
 (0.01 secs, 1,632,784 bytes)

divergenceCheck :: Maybe String

If this returns anything other than:

 Just "It didn't diverge!"

Then we have a bug.

The bug it's designed to find is: either tabulateFold or indexFold is non-productive until reaching the end of the input. In this case, the input is unlimited while the Fold can return upon reaching the first element.

And.. it's not returning..

 λ> divergenceCheck
 ^CInterrupted.

Well we have a bug..

Now I'm wondering whether that's a feature.. of the strict implementaton of fold..

It's a feature, regular folds also diverge:

 λ> F.fold F.head [1..]
 ^CInterrupted.

Even using generalizeMaybe diverges:

 λ> F.foldM (generalizeMaybe F.head) [1..] :: MaybeT Identity Int
 MaybeT (Identity ^CInterrupted.

Orphan instances

Distributive (Fold a)

Instance details

Methods

distribute :: Functor f => f (Fold a a0) -> Fold a (f a0)

collect :: Functor f => (a0 -> Fold a b) -> f a0 -> Fold a (f b)

distributeM :: Monad m => m (Fold a a0) -> Fold a (m a0)

collectM :: Monad m => (a0 -> Fold a b) -> m a0 -> Fold a (m b)

Representable (Fold a)

Instance details

Associated Types

type Rep (Fold a) :: *

Methods

tabulate :: (Rep (Fold a) -> a0) -> Fold a a0

index :: Fold a a0 -> Rep (Fold a) -> a0