Data.Foldable.Monadic.NonEmpty

Description

This module provides a class for FoldableM, restricted to data-types that are non-empty up to the Monad being non-empty.

Documentation

class FoldableM m t => FoldableM1 m t | t -> m where

FoldableM for non-empty containers

Minimal complete definition

pureFold1, impureFold1

Methods

pureFold1 :: Fold1 a b -> t a -> m b

impureFold1 :: FoldM1 m a b -> t a -> m b

pureFold1MaybeT :: FoldableM m t => Fold1 a b -> t a -> MaybeT m b

Return empty unless the container being folded over is non-empty. Otherwise behaves like pureFold.

impureFold1MaybeT :: FoldableM m t => FoldM1 m a b -> t a -> MaybeT m b

Return empty unless the container being folded over is non-empty. Otherwise behaves like impureFold.